Add problem_change support to SolverManager
To support adding/removing items from the list in problem changes,
the wrapper types of @planning_entity_collection /
@problem_fact_collection changed from array to List.
Arrays are copied and thus changes to the list are not reflected
in the array. @value_range_provider are still arrays, since
raw Python types (ex: datetime.date) cannot be used in PythonList.