List available subpackages in optapy.config, validate collection decorators
Now that Lists are being used instead of Arrays so the Java Collections
will match the Python list after a problem fact change, an illegal
object will be put on the stack if any Python class is return by
the list get. To avoid this, we validate the type passed to
@planning_entity_collection and @problem_fact_collection to
ensure it is a planning entity or problem fact.

Use __all__ in optapy.config so IDE can list all packages.