WIP: Started work on Python -> Java translator
Architecture:

Code is translated when the ConstraintProvider is constructed
in its defineConstraints method. Each function passed to
Joiners, filter, Collectors, etc. get translated to a Java
method. This prevents needing to analyze/translate the defineConstraints
function, since we only translate what we need (when it passed
to ConstraintFactory/Joiners/ConstraintCollectors).

Python side, use the 'dis' module of Python to get the bytecode.
There are several attributes of the code object we need:

co_names: list of names used in the function
co_varnames: list of local variables used in the function
co_constants: list of constants used in the function

co_constants can pretty much contain anything, so future work
need to be done to translate the python value to their java
counterpart.

Python, like Java, is a Stack based language. As such, their bytecode
is similar.

Implemented:

- NOP
- POP_TOP
- ROT_TWO
- ROT_THREE
- ROT_FOUR
- DUP_TOP
- DUP_TOP_TWO
- RETURN_VALUE
- LOAD_CONST
- LOAD_NAME
- BUILD_TUPLE
- BUILD_LIST
- BUILD_SET
- COMPARE_OP
- LOAD_FAST
- STORE_FAST