Support numeric operations in Python translator
Because everything an Object in Python, Numbers are Objects.
Moveover, the same Bytecode instruction that corresponds to
add, subtract, etc. also works on objects via dunder methods.

(i.e. if type(a) has attribute __add__, a + b = type(a).__add__(a,b))

Further, if you inspect Python numbers, you see they also have these
dunder methods.

Thus, it now required to define the interface for all "Python" object.
Fudamentally, two things are required:

- a way to get attributes
- a way to store attributes

(i.e. a Map).

Wrappers of List, Function, Long and Double were created that implement
this interface, and for Long and Double, provide dunder methods for
their operations.

A Copy-On-Write map is used for the default dictionaries of Long,
Double, List, (and all built-in types), so modification of the map
does not affect new instances.

Newly supported bytecode instructions:

- BINARY_POWER
- BINARY_MULTIPLY
- BINARY_MATRIX_MULTIPLY (not implemented in types)
- BINARY_FLOOR_DIVIDE
- BINARY_TRUE_DIVIDE
- BINARY_MODULO
- BINARY_ADD
- BINARY_SUBSTRACT
- BINARY_SUBSCR (not implemented in types)
- BINARY_LSHIFT
- BINARY_RSHIFT
- BINARY_AND
- BINARY_XOR
- BINARY_OR
- INPLACE_POWER
- INPLACE_MULTIPLY
- INPLACE_MATRIX_MULTIPLY (not implemented in types)
- INPLACE_FLOOR_DIVIDE
- INPLACE_TRUE_DIVIDE
- INPLACE_MODULO
- INPLACE_ADD
- INPLACE_SUBSTRACT
- INPLACE_LSHIFT
- INPLACE_RSHIFT
- INPLACE_AND
- INPLACE_XOR
- INPLACE_OR
- COMPARE_OP