Add solver_config_create_from_xml_file function
The class name to use in *Class elements is the
module + class/function name. If the module is __main__,
then only the class name to be supplied. Example:

```python
from my.domain import MyEntity # my.domain.MyEntity

class MySolution: # MySolution
    ...

if __name__ == '__main__':
    solver_config =
    solver_config_create_from_xml_file(pathlib.Path('solverConfig.xml'))
```

The code works by creating a new ClassLoader from the current alias
map, which allows multiple classes with the same module + name to
exist, of which the latest defined is used (think notebooks and terminal).