Add type infomation to PythonLikeObject; move dunder impl to type
For performance optimizations, CPython does not check for dunder
methods on objects and instead go directly to the type. Thus,
the directories that contain them for PythonInteger and co. were
moved from the object themselves to their newly defined types.

Additionally, made buildList and buildTuple create a PythonLikeList,
and fixed a bug in loadConstant.