Add support for iterators and jump absolute; fix unary operators
In Python, iterators indicate they are exhausted by raising
StopIteration. To implement this in bytecode, we put the
iterator loop into a try block, and catch NoSuchElementException
(which is Java's StopIteration exception that iterators should
throw).

There was a bug with the unary operators bytecode; it should
use dup_x1 instead of dup_x2 since there is only one operand on
the stack.

Newly Implemented Opcodes:

- JUMP_ABSOLUTE
- FOR_ITER
- GET_ITER