Add support for cell variables, free variables, inner functions
Unlike Java, Python's inner functions can observe changes to
the outer function variables. Python accomplishes this with
cells, which functions a lot like AtomicReference.

Because Python function support some...features...not
supported by Java functions, a slightly different class
(with new instance fields) is created when the requested
interface is PythonLikeFunction. It has the following
instance fields:

- __closure__, to store cell and free variables
- __qualname__, to store the qualified name of the function
- __annotations__, to store annotations on the function
- __kwdefaults__, defaults for keyword only arguments
- __defaults__, default for positional only/positional or keyword arguments

Additionally, PythonCompiledFunction now has
co_argcount and co_kwonlyargcount so the arguments
can be extracted for a PythonLikeFunction.

Once a function is compiled, its class can be loaded
as a constant (which converts it to a PythonCode that
just holds a reference to the class). That class then
can be used with MAKE_FUNCTION to create an inner
function.

Newly supported Opcodes:

- MAKE_FUNCTION
- LOAD_CLOSURE
- LOAD_DEREF
- STORE_DEREF
- DELETE_DEREF