Add support for exception handling opcodes
The SETUP_FINALLY declares a try block, and has a reference
to the first exception handler for that try block. The
exception handlers jump to the next one if the Exception
does not match with the JUMP_IF_NOT_EXC_MATCH instruction.
RERAISE raises an exception on TOS. RAISE_VARARGS is used
to support three forms of raise:

- raise (re-raise the last exception)
- raise TypeError (raise an exception (either from a type or instance))
- raise TypeError from e (raise an exception with a cause)

Added support for checking subclasses on types in order to implement
the JUMP_IF_NOT_EXC_MATCH. <clinit> was not being called before
<init> for some tests for PythonType, leading to a NPE. To fix
that, BASE_TYPE creation been move to a static method that is
called in <init>.

Newly supported opcodes:

- LOAD_ASSERTION_ERROR
- POP_BLOCK
- POP_EXCEPT
- RERAISE
- SETUP_FINALLY
- RAISE_VARARGS