Make optapy import translator from javapython
Because it is impossible to use directories below the project
directory in setup.py, optapy-core setup.py files + tests
were moved to project root. This allows building the python to
java bytecode translator as a submodule of the optapy project,
while still allowing it to be built independently of optapy.

Additionally, a bug fix for subclasses. Previous, a field to
store instance fields in the Java Python interpreter was created
for the subclass and all its parent class. Now, it is only created
for the top parent class (which the subclasses access).