Only use Python setters on clone if possible; translate all CS methods
If everything in the domain and constraints can be translated to
Python, the setters on the generated wrapper classes will not go
to Python until cloning occurs, massively increasing performance.
To accomplish this, a new method, `forceUpdate`, was added to
the `PythonObject` interface. The `forceUpdate` method forces
the Java wrapper to update the Python object. On the Planning
Solution, this call force update on all the Planning Entities.
For Planning Entities, call the setter of all its planning
variables (geninue and shadow). For Problem Facts, forceUpdate
does nothing, since Problem Facts do not get updated.

Wrappers were created for all Constraint Streams classes. This
allows us to translate functions in filters, groupBys and
penalize functions. Additionally, it gives us additional
type infomation about the content of the stream, allowing
for potential future optimization.

Because PythonObject are automatically proxy in Python,
The convert method does not see it as a PythonLikeObject.
To overcome this, the translator check if the
object to convert was generated by optapy, and if so,
return the id. `getAttribute` then can check if the
result returned was int, and if so, look it up in the
id to object map.

Added default implementations for `__hash__`, `__eq__`,
and `__neq__` so `==`, `!=` and `hash` will work as
expected in the Java Python Interpreter.