Add dunder methods used by bytecode in proxies
For optimization reasons, the Python interpreter
check the types and ignore the object attributes
when looking for dunder methods. Additionally,
PythonComparable was not proxy, causing errors
when it was accessed. Thus to allow subscripting,
addition, and other operators on Python object,
the Proxy class need to implement all dunder
methods checked by the bytecode (with a lookup
on the delegate).