Remove memory leak of old best solutions in cloning
Before, all clones of a solution were kept in memory
until solving is completed. However, this is not
neccessary, since OptaPlanner only need the current
best solution and the current working solution
(i.e. the past two best solution). Thus it was
leaking unneccessary memory.

We track what run a problem on with a
new attribute: _optapy_solver_run_id. Because
the same problem could be used in multiple
solver runs, the problem is shallowly cloned
on calls to Solver/SolverManagers so
it will not conflict with existing
_optapy_solver_run_id.

This new attribute eliminated the need for the
ref_id_to_solver_run_id map, and simplify
setup and cleanup logic.