Refactor opcode logic; get stack type metadata for each opcode
In order to support multiple Python version (and to split
the 100+ cases switch into multiple more managable ones),
the logic for implementing opcodes is moved to the Opcode
interface. The inner class PythonInstruction.OpCode changed
to OpcodeIdentifier (and the argument to the enum was removed
because it was not used).

A flow graph is created from the opcodes. This flow graph is
used to determine types of local variables, cell variables,
and stack elements.

The flow graph is created using the following procedure:

1. Create basic blocks from the opcodes. A basic block is
   a block of code that does not jump into the middle of
   itself (it may jump to the beginning) and does not
   have an absolute jump (or an instruction that terminates
   the function) instruction in the middle (it may be
   at the end).

2. Using the branch infomation from the opcodes, create the
   edges of the graph (stored in basicBlockToJumpSourcesMap).

To get type metadata, the following procedure is performed:

1. Set the metadata for the first instruction to the empty stack

2. For each basic block:

2.1  For each opcode in the basic block:

2.1.1  get the metadata for the instruction

2.1.2  For each branch of the opcode (most opcode has one branch: the
       next instruction)

2.1.2.1  get the metadata if that branch is taken

2.1.2.2  if the branch goes to another basic block, update the jump
         source metadata

2.1.2.3  unify the metadata of the branch with the current metadata of
         the branch (simply setting it if it is null)

3. Repeat 2, but before 2.1, unify the first instruction metadata
   with all the jump sources metadata

4. Repeat 3 while unifying changes the metadata

The stack metadata will be used later to optimize bytecode generation.
For instance, if TOS and TOS1 are PythonInteger, we can use an
invokeVirtual instead of a complex series of attribute lookup,
which should massively improve performance.