Unproxy arguments to dunder methods
For certain generated methods, such as those
generated by dataclasses, may use is comparisions
which do not invoke any dunder method and thus fail
for proxies. Thus, before passing the arguments to
the dunder method, they must be unproxied so the
dunder method return the correct result.

This was a particular issue for dataclasses with eq=False,
since they use "is" comparsions, causing a uncomparable
exception if they were used as a tuple in a groupby key
(and just overall being incorrect).