Use and translate the function's globals dict
Before, a seperate globals map was used in
CPythonBackedPythonInterpreter to keep track of globals.
However, this is not suffice: functions in different modules
can have different global maps, and the seperate globals map
did not have any globals defined.

To fix this, we take the globals dict from the function's
__globals__ attribute and translate it into a Map which
is then stored as a static field of the compiled function.
During bytecode translation, if we detect a global that
could not be translated is used, we throw an exception
since calling the function in CPython will be faster
due to the marshalling cost involved with going to
CPython and back.

PythonObjectWrapper was updated to include a reference
to the CPython type of the object, which allows it to
behave correctly when interacting with other objects.

A missing attribute will now throw an AttributeError
instead of an UnsupportedOperation exception.