Wrap constraint streams in Python delegate; make get_class idempotent on Class instances
Wrapped the constraint stream API in a Python delegate, which allows
us to wrap the functions passed to groupBy (and other constraint stream
operations) so they return Java compatiable objects.

get_class was not idempotent on Class instances (return Object.class)
instead, which caused some problem when it was used in the constraint
stream delegate. It is now idempotent.