Determine known method for dunder operation from stack types; formatting
PythonLikeType now keep a reference to the Java class it represent,
which is needed for typecasting (so no verify errors occur when
running the bytecode). Additionally, it has a new methods:
addMethod(...), which add a known method overload, and
getMethodType(String), which get the overloads corresponding
to the given method name.

A overload (stored as a PythonFunctionSignature in
PythonKnownFunctionType), stores a MethodDescriptor
that knows how to invoke the method, as well as the
parameter and return types (which is used when determining
which overload to invoke and what the new TOS element
will be).

Added known methods for PythonFloat and PythonInteger.