Support translating Python classes to Java classes
To convert a Python class to a Java class, we do the following:

1. Get all the methods defined on the Python class
    - Python has three types of methods:
        - instance, take self as first parameter
        - static, can be performed without an instance
        - class, can be performed without an instance, take type as
          first parameter
2. Perform typeflow analysis on instance methods to determine what
   attributes are accessed on self
3. Use the set of attributes accessed on self to generate field
   descriptors for that class (allowing for opcode optimization)
4. Create static fields to hold implementations of each method
5. Create Java methods that call the static field implementation
6. Override getAttribute, deleteAttribute and setAttribute to
   get/set/"delete" (set to null) the field
7. Generate bytecode for each method and set the corresponding
   static holder field to the generated instance.

To accomplish this, generating bytecode takes MethodDescriptor
(which uses ASM Types) instead of a Java interface class
(which uses classes). This is to allow bytecode generation
when a class does not exist yet.

Additionally, StackMetadata now tracks ValueSources, which,
as the name implies, track the initial source of a value.
This is used to identify if "self" is used in
GET_ATTR, STORE_ATTR, and DELETE_ATTR opcodes.

Generated classes (and some non-generated classes) store their
PythonLikeType in the $TYPE static attribute, which allows
getPythonLikeType(Class) to create the initial ValueSources
of parameters.