Use generics when generating bytecodes for methods; class translation bugfixes
Now that we are using typeflow analysis, it is beneficial to pass
type infomation when generating the Java implementation of the
methods. Joiners were changed to return a data object containing
the property function(s), so they can be translated when
type infomation is available (at the Constraint Stream
join, ifExist, etc. call sites).

A new class OptaPyObjectReference which wraps a long
is used when translating ids, since when wrapping a list of
PythonLikeObject, you can only add PythonLikeObject to
the list.

Fix various bugs with class translation:

- The hashcode switch jump back to the start of the current
  case instead of to the next case
- Parameters were not named/too many parameters, causing
  problems decompiling the code
- Do not generate bytecode for inherited methods
  (for example, object.__eq__)
- Use actual superclass of type