Fix Custom Variable support
There were two issues with custom variables that caused
them to be broken with bytecode generation

1. Changes Java side do not propagate to Python. The long
term fix is to check if any domain class contains or
references python code that is not translated when
the constraint provider is generated (thus forcing
setters to go back to Python). The short term fix is
to force update the entire solution; in theory,
it might be enough to only force update the entity
that change, but custom variable listeners can
be complex and reference other entities.

2. Changes in the Python side do not propagate to Java.
This is fixed by created a wrapper for ScoreDirector,
which updates the Java side of the variable whenever
afterVariableChange is called.