Automatically cast flattenLast and ConstraintCollectors arguments
Created a ConstraintCollectors delegate similar to the Joiners,
which automatically cast its arguments to the correct types.
Flatten last need to use a Java collection type for it to properly
unpack its arguments. The JConversions for ToIntFunction can be
removed since everything that uses them now use the above
wrapping methods.