Add datetime and overload generation support
The datetime module is heavily used by domains, and their
types are builtin, meaning we need to provide our own implementation.
Thankfully, they are a lot like Java LocalDateTime, LocalDate,
LocalTime, and Duration, which we delegate to.

In the case when the type system cannot determine if TOS
has a particular method, the generic code is used, which
gets a method from the object's type. This is difficult for
overloaded methods, as there are multiple candidates. To make it
easier to create (and less error-prone), an class generator was
created which create the Dispatch classes (which takes the list and
map, determine the best candidate, and call it).

This commit contains some formatting changes that were missed.