Add type infomation to generated class fields
Type infomation is obtained from the class' __annotations__
dict, if it exists. A condition in the constructor was inverted in PythonWrapperGenerator
(it suppose to  checks if an object can be assigned to a PythonLikeObject).
Because None cannot be assigned to fields that are not
PythonLikeObject/PythonNone, it is converted to null when setting
fields.

In Python, generics types are not the same as the base type
(i.e. list[int] is not list), so we need to erase the generics
args when resolving annotations (so we can look up the type
in the built type directory).