Optimize CALL_FUNCTION and CALL_FUNCTION_KW
Optimize CALL_FUNCTION and CALL_FUNCTION_KW in
cases where the function that will be called is known.
To handle default arguments and keywords, a new class
is generated for each signature, signature$$Defaults,
which is defined by PythonDefaultArgumentImplementor.

$$Defaults hold the constants used as the default arguments,
and has a method to "build up" the argument list the same
way CALL_FUNCTION_KW does it (construct with keyword tuple,
then add keyword arguments, then positional).

Moved switch logic out of ClassTranslator and into new class,
BytecodeSwitchImplementor, which allow it to be reused elsewhere.

Modified the __get__ function for functions to return a
BoundPythonLikeFunction, which will allow the code to
extract the instance the function is bound to.