Optimize comparisions and equality
Changed PythonLikeComparable to an interface that
define lessThan, greaterThan, etc. methods. Modified
its setup to use addMethod(...) instead of modifying
the dir. Modified the base type to define equals and
notEquals methods.

Because the TYPE static fields often refer to each other
when they are being defined, they may sometime be null
despite being static final. To fix this, when types
are being defined, a static getter method is used
to get the type.

Created a new MethodType, STATIC_AS_VIRTUAL, which allows
a static method to be used as a virtual method
(example: for Object.equals).