Reuse existing references when cloning PythonLikeObject fields
There are many cases where identity equality is required
(for instance, a map with a key that does not define a hashCode).
Thus, when cloning PythonLikeObject fields, we need to reuse
existing references. To help with this, convert_to_java_python_like_object
now take an optional Map parameter to return existing references. A
global map cannot be used, as that will cause a memory leak.
Additionally, PythonLikeObject based on Python classes can
now read from/write to their backing CPythonReference with
special method.

There is a special interaction with cloning: we want the
PythonLikeObject to reuse our cloned objects, and not
create new ones (which will cause cloning corruption).
However, the OptaPy reference map contains mappings
from old (before clone) ids to the (Python) clone objects.
We want to have a map from the clone objects ids
to the (Java) clone objects. To accomblish this, before
setting the PythonLikeObject's fields, we visit all the
ids accessible from the planning solution, adding them
to the reference map, and pass the reference map to the
cloner.