Start a Python thread to catch SolverManager exceptions
Exceptions from Java threads are hidden if there is no Python
thread attached to them. This is an issue with SolverManager,
since it create threads in solve and solveAndListen, meaning
we don't get to see their exceptions. To fix it, we create
a new Python thread that joins the returned SolverJob, thus
catching any exception the Java thread might throw.

Fixes #91