Use non-backing Java list if Python setters are ignored; mirror reference map to prevent use-after-free
To increase performance when using problem facts collections, planning entity
collections, and planning list variable, a list that is NOT backed by
CPython is used if Python setters are ignored (i.e.
PythonWrapperGenerator.NONE_PYTHON_SETTER).

Previously, the Java Python Interpreter reused the OptaPy reference map
to store its mapping from Python Object id to Java Objects. However,
by doing so, it also pollutes the mapping with the ids for
planning list variables and planning scores, which are temporary
and can be freed when a solution is cloned. This leads to cases
where, after an object is cloned, it has the same id as a freed
planning score/planning list variable. This cause a use-after-free
that typically results in a exception (wrong type). To fix it,
instead of giving the Java Python Interpreter the OptaPy reference map,
we give it a mirror to it. The mirror can add extra entries without
affecting the original reference map. When the solution is cloned,
a new mirror is passed, removing the old (and potentially freed)
entries.

Because the lists are no longer guarateed to be backed to Python,
the ProblemChangeDirector now clears the reference map and
reset all the fields (which repopulate the reference map)
on the planning solution before returning to Java.