Translate classes in global; fix try-except handlers; fix self-type static attributes
Everything in a method's globals will be (attempted to be translated).
It will not be translated if it is built-in or c-native. Since the
function's globals may contain the function itself, a key-set is kept
for each global map, so a function is not translated in an infinite
loop.

Try-Except handlers actually have 6 items on the stack:
the source instruction, the stack size before the try block,
the label to the handler, traceback, the exception, and the
exception type. The bottom three stack values are not used
by us (since Java take care of it for us), but the Java
stack size at a particular instruction MUST match the Python
stack size, so we push three values (None, Stack Size, None)
before the traceback, exception and exception type. Additionally,
Python expect the stack to contain all values from before the
try-block, unlike Java, where the stack only contain the exception.
So the stack is stored in new Java locals and restored before
entering a try block.

Self-type static attributes (example: values in an enum) were
previously translated before the class was created. This is an issue
since it will become a CPythonObjectWrapper. To fix, they are treated
seperately from the other attributes (which are translated before
going to Java) and are translated after the class is created.