Make PythonList use cached list for read operations
For read operations (get, size, isEmpty, etc.),
PythonList will use a cached Java List of values.
This removes going to Python for the majority of
operations and grants a massive speed boost.

Write operations (set, add, etc.) perform their
operation on both the cached Java List and Python
List.

We are unconcerned with changes that happen in
Python, since PythonList is used only for
PlanningEntityCollection, ProblemFactCollection
and PlanningListVariable, and if those collections
change OUTSIDE of OptaPlanner, that will cause
score corruption.