Fix off-by-one-error in comprehension opcodes; store/restore stack before/after FOR_ITER try...catch block; Store globals + constants in same Map; free unused references after cloning
There was an off-by-one-error in comprehension opcodes. For instance,
LIST_APPEND(arg=3) modifies TOS[4] instead of TOS[3]. This
contradicts dis docs "Calls list.append(TOS1[-i], TOS)", but matches
behavior in Python.

Due to how the JVM works, the stack is empty after a try...catch
block. Due to how Python works, it expects the stack after
a FOR_ITER loop to have the same contents as before the loop.
Also due to how Python works, we need to use a try...catch
block to catch a StopIteration exception. The previous
implementation did not store/restore the stack before/after
the loop, leading to inconsistent stack sizes
(mainly for comprehensions/multiple loops). The fix is to
store the stack in local variables before the loop, and
restore the stack after the loop.

Globals and constants now use a shared static map. Additionally,
when solving started, it create a new Mirror map of this shared
static map. This is so constants/globals use the same
Java instances, without polluting the shared map with temporary
objects.

Unused references after cloning (i.e. old entities) are now removed,
preventing a memory leak.

Unrelated, but super helpful: discovered CheckMethodAdapter
from asm-util, which tells you HELPFUL ERROR MESSAGES when
there are errors in the bytecode (instead of NullPointerExceptions).