Add support for modules, IMPORT_NAME and IMPORT_FROM opcodes
Modules are now translated; this is done by taking their
attributes and converting them to Java.

Importing modules in Python can have side-effects, so we cannot
"import" them eagerly during compilation. Thus, when we import,
we first check if the module is in our cache. If so, return that.
Otherwise we call the Python __import__ function to perform the
import for us and cache it.

IMPORT_FROM is basically a DUP_TOP followed by a LOAD_ATTR.

IMPORT_STAR is not supported, since it is a syntax error in
Python functions as it is only allowed in the module level
(which we do not deal with).