Ban pytest module; check if class is defined in banned module
It appears an infinite loop occurs when translating pytest's
TerminalOutput class. To allow testing, pytest module is now
banned (so classes defined in pytest will not be translated).

Add a check to see if a class was defined in a banned module.
If it is, we do not want to translate it (since banned modules
have classes/functions that cannot be translated).