Add generator support; correctly handle nested try's
Python generators are coroutines: a function that
yield a value, wait, and then resume (until it
yields another value). The generator resumes with
the same stack and locals as when it was yielded
(with an exception: TOS is replaced with the sent
value, or None if no value was sent). Java does
not have a corresponding mechanism for coroutines;
the closest is an Iterator class.

To implement coroutines in Java, when the function
yields a value, we first need to store the stack
into an instance field (so it can be restored
when the coroutine resumes). Additionally, we
need to record where we yielded at, so we can
jump to the correct position when resumed.

When resuming a generator, we push the items
from our instance field onto the stack, along
with (the potentially None) sent value, then
jump to the correct postion. Before resuming
the code, we check if user wanted to throw
a value (via the generator's throw method).
If the generator thrownValue field is not null,
it is thrown (and potentially handled by the
generator own except handlers, or rethrown if
not handled.

To allow generators to be used in for loops,
they implement the Iterator interface.
hasNext and next checks if it needs to resume
the coroutine (i.e. has next returned the last yielded value).
hasNext returns true iff and generator did not end at a return
statement, and next return the yielded value
(when the generator hits a return statement, it throws
StopIteration with the returned value).

To save locals, a subclass of LocalVariableHelper
was created. Instead of saving/reading locals from slots,
the subclass saves/reads locals from instance fields of
the generator (creating them if absent for temp variables).
In order to support this, opcodes that used temp variables
were modified to use the LocalVariableHelper methods for
reading/writing locals (instead of directly doing
methodVisitor.visitVarInst(...)).

Nested try's were not implemented correctly; when try's
were nested, the outermost try is executed, instead of
the inner try. In Python, most try-blocks are actually
two try-blocks: one for exceptions in the try block,
another for exceptions in the except/finally handlers
(which rethrow the exception). To fix it, we use an
MethodVisitorAdapter which reorders the except handlers
so the shortest (inner-most) one is first.

When compiling a function, we check if it is a generator
(i.e. has yield opcodes), and if so, we compile it using
PythonGeneratorTranslator instead of the normal path.

Newly supported opcodes:

- YIELD_VALUE