Add support for yield from expressions in translator
The "yield from subgenerator" syntax in Python causes
a generator to delegate next, send and throw calls to the
subgenerator, returning the subgenerator's values until it
is exhausted. It is roughly the same as the following Python code:

    try:
        while True:
            try:
                sent_value = (yield)
                if sent_value is not None:
                    yield subgenerator.send(sent_value)
                else:
                    next(subgenerator)
            except Exception as e:
                subgenerator.throw(e)
    except StopIteration:
        pass

If subgenerator is an iterable (i.e. not a generator), then
sending a value raises either AttributeError or a TypeError,
and throwing a value immediately raises it.

It is implemented by adding a yieldFromIterator field to
generator classes, and delegating its next(), send(), and
throw() calls to it until StopIteration is raised.

GET_YIELD_FROM_ITER is basically iter(TOS), but it leaves
TOS alone if it is a generator or coroutine.

GEN_START was also implemented; despite having a stack effect
of 0 reported via StackMetadata, it actually pops TOS.
Generator functions always have the sent value (or None) when
starting/resuming. Since it impossible to send a value until
the first yield, the stack at the start of a generator function
is [None], and after this opcode, it changes to [].

This change apparently fixed the broken inverse_relation test.

Newly supported opcodes:

- YIELD_FROM
- GET_YIELD_FROM_ITER
- GEN_START