Add support for Python's with blocks
Python with blocks are equalivant to Java's
try-with-resources; this code in Python:

    with context as a:
        ...

is equalivant to the following code in Java:

    try (Context cxt = context.get()) {
        ...
    }

Java's Closable close method is mapped to __exit__ in Python,
and the assignment happens using the __enter__ method in Python.

The SETUP_WITH loads __exit__ and the result of __enter__,
then create a try...finally block. WITH_EXCEPT_START calls
__exit__ (which in Python 3.9, is at the 7th position on the
stack) with the exception info, leaving the stack unchanged
(except for pushing the result of __exit__ to the top of the
stack).

To help debug errors in type flow analysis, The location of
the block that cannot be unified along with the bytecode of
the function is now displayed when a type flow analysis error
occurs.

Instruction argument was replaced with exceptHandlerLocation,
so try code can be reused in multiple places.

Newly supported opcodes:

- SETUP_WITH
- WITH_EXCEPT_START