Translate Python Code Objects; Initialize closures for parameters; Fix Python 3.9 jump target args
The Python translator will now translate code objects, which allows
comprehensions and inner functions to be translated.

When parameters are used as cell variables, the cells are
initialized with the parameter (i.e. without STORE_DEREF).

Python 3.10 has an undocumented difference to Python 3.9:
the argument of instructions that can potentially jump
are halved (in Python 3.6+, each instruction takes two bytes,
so it changed from byte position (which is always a multiple of 2),
to instruction position).

With this commit, the target of jumps in the translator is always
absolute (and can be obtain via JumpUtils). This make it easier
to reason about the actual jump location when writing the Java
bytecode. This differs from Python, where the jump argument
can indicate either a relative or absolute jump (and may change
between versions, even for the same instruction). Since our
internal Opcodes now only use absolute jump locations, the
internal JumpForward opcode is removed and replaced with
a JumpAbsolute.

To help with managing differences between Python versions,
a value class is used that breaks down sys.hexversion into
its parts. This modifies the Opcode interface to take that
object instead of the hex version.

Tox is now used with the python translator to ensure
the functions are translated corrected when running in
the CPython environment. It is slower than running the
Java tests, but will catch bugs related to undocumented
changes to dis and bugs caused by an incorrect understanding
of the opcode.