Ignore dead Python bytecode, populate modules lazily, add some builtins
CPython can generate dead bytecode. The dead bytecode confuses
FlowGraph, since it needs to know the stack metadata state
for each instruction, which it cannot generate if the code
is dead (since dead code do not have a basic block pointing
to it). When dead code is encountered, it is assigned a
DEAD_CODE stackmeta, which acts as an identity element when
unifying (in particular, DEAD_CODE unify x = x unify DEAD_CODE = x).

Previously, modules were populated eagerly. This is expensive,
sometimes taking several seconds. Additionally, it can lead
to redefining builtins, breaking code. Now modules are
populated lazily (i.e. a module attribute is loaded
when it is first requested; and module attributes
are loaded seperately instead of all at once).

Fix an infinite_loop_bug in erase_generic_args; in certain
cases, typing.get_origin(x) == x, causing an infinite loop.
Since the origin of a type should not have an origin
(other than possibly itself), the while loop was changed
to an if.

Fix a bool conversion bug; getValue() (which is inherited from
PythonNumber, and thus return a number) was used when converting
from Java to Python. Now the Python boolean value is obtained
by checking equality with PythonBoolean.TRUE.

Previously, when a python type was encountered that did not
have a corresponding Java class, a CPythonType was returned.
Now, it attempts to translate the type.

Add a constructor for type, which is typically used to get
the type of a python object (and sometimes used to create
a new class).

Add a constructor for int, which calculate the int value
of an object.

Add a constructor for bool, which calculate if an object
is considered "true" in boolean contexts.

Add a placeholder for the builtin complex type.

Fix a bug in DELETE_SUBSCR; it should ignore the result
of the delete function (previously, it kept it on the stack).
Additionally, it was misspelled as DEL_SUBSCR.

Add CHECKCAST instructions for function calls (CALL_METHOD,
CALL_FUNCTION, CALL_FUNCTION_KW) and when trying to throw
exceptions (RERAISE, RAISE with cause) to prevent
VerifyErrors when loading the class.

Add class static variables to the global shared instances map.
Previously, they were not added, causing two different Java
instance to exist for the same Python instance, causing bugs.

Add support for atypical Python classes (example: extending int,
not having any super classes, virtual methods without a self arg).

Add support for some builtins. Simpify code for unary dunder builtins
by creating a class to hold the common code.

Currently supported builtins:

- abs
- all
- any
- bool (type)
- dict (type)
- filter
- float (type)
- int (type)
- iter (1-arg only)
- len
- list (type)
- min
- max
- next
- repr
- set (type)
- str (type)
- tuple (type)
- type (type)
- print